feat(config): executables for examples compatible with new config struct

+++ src/main_impl.hpp
@@ -208,7 +208,7 @@ inline Observer auto create_gridboxes_observer(const unsigned int interval, Data
 template <typename Store>
 inline Observer auto create_observer(const Config &config, const Timesteps &tsteps,
                                      Dataset<Store> &dataset) {
-  const auto obsstep = (unsigned int)tsteps.get_obsstep();
+  const auto obsstep = tsteps.get_obsstep();
   const auto maxchunk = config.get_maxchunk();
   const auto ngbxs = config.get_ngbxs();
